requirements.txt
# Core dependencies
kafka-python>=2.0.2
praw>=7.7.1
pandas>=1.5.3
numpy>=1.24.3
scikit-learn>=1.0.2
joblib>=1.2.0

# Database
sqlalchemy>=1.4.46
psycopg2-binary>=2.9.6

# Dashboard
streamlit>=1.22.0
plotly>=5.13.1

# Utils
python-dotenv>=1.0.0

# Testing
pytest>=7.3.1
pytest-cov>=4.0.0